backend/requirements.txt
fastapi
langchain
google-generativeai
tavily-python
pydantic
uvicorn
black
ruff
mypy
pytest
python-dotenv
langchain-google-genai
pillow